include versions of code used in Debrief
