+++ src/DopplerCurve.java
@@ -1,3 +1,4 @@
+
 import java.util.ArrayList;
 import java.util.Collection;
 
@@ -106,7 +107,6 @@ public class DopplerCurve implements IDopplerCurve
   private final double[] _modelParameters;
 
   private final Normaliser _timeNormaliser;
-  private final Normaliser _freqNormaliser;
 
   public DopplerCurve(final ArrayList<Long> times, final ArrayList<Double> freqs)
   {
@@ -128,7 +128,7 @@ public class DopplerCurve implements IDopplerCurve
       dTimes.add((double) t);
     }
     _timeNormaliser = new Normaliser(dTimes, false);
-    _freqNormaliser = new Normaliser(freqs, true);
+    final Normaliser freqNormaliser = new Normaliser(freqs, true);
 
     final int sampleCount = times.size();
 
@@ -138,7 +138,7 @@ public class DopplerCurve implements IDopplerCurve
     for (int i = 0; i < sampleCount; i++)
     {
       double time = _timeNormaliser.normalise(dTimes.get(i));
-      double freq = _freqNormaliser.normalise(freqs.get(i));
+      double freq = freqNormaliser.normalise(freqs.get(i));
       obs.add(time, freq);
       System.out.println(time + ", " + freq);
     }
@@ -158,7 +158,7 @@ public class DopplerCurve implements IDopplerCurve
 
     // use bisection solver to find the zero crossing point of derivative
     BisectionSolver bs = new BisectionSolver(1.0e-12, 1.0e-8);
-    double root = bs.solve(1000000, derivativeFunc, 0, 1, 0.5);
+    double root = bs.solve(1000000, derivativeFunc, 0.01, 1, 0.5);
 
     // and store the equation parameters
     _modelParameters = coeff;
