+++ src/DopplerCurveFinMath.java
@@ -1,3 +1,4 @@
+
 import java.util.ArrayList;
 
 import net.finmath.optimizer.LevenbergMarquardt;
