+++ src/IDopplerCurve.java
@@ -1,3 +1,4 @@
+
 public interface IDopplerCurve
 {
