+++ src/Normaliser.java
@@ -1,3 +1,4 @@
+
 import java.util.ArrayList;
 
 import junit.framework.TestCase;
